Fix get_location for opts in groups
get_location defines the type of its ``group`` parameter as a str,
but it then passes that directly into _do_get, which expects an
OptGroup for its parameter. When it tries to reference group.name
it blows up because a str doesn't have a .name attr.

In the interest of not changing either API, this change coerces the
group str into an OptGroup.

Closes-Bug: 1807230
Change-Id: Iae547cbd63059fa19691a7f7fe398b148effd177