Document precedence of config sources
Previously the precedence details for the various sources oslo.config
can consume were scattered around the developer documentation (for
example in [0]). Since this is relevant for users too, we need to
include it in the user-focused documentation.

This collects all of the precedence information I am aware of and puts
it on the documentation page for --config-file, --config-dir, and the
sample source driver options.

It also clarifies the file and dir precedence in the developer docs since
previously that was only available in the API docstring.

Change-Id: I41d8ac0f33ea60aea8ba5b150e53d7eafa323457
0: https://github.com/openstack/oslo.config/blob/0bc22839b09fcd7ae2de4f7fe5d63ac055fd0a78/oslo_config/cfg.py#L2095
Closes-Bug: 1825859