Add a Quick Start tutorial
As a user completely new to oslo_config (well, I had only experienced it
in the already-implemented context of the nova project), I would like to
have had a reference document that provided a step-by-step guide to
writing a simple test script I could use to noodle around.  It might
have gone something like this.

Change-Id: I02ff56f47ac8b93e0ee37eb55b53552059e4da27