Deprecate ConfigFilter
There are no existing users of this class[0], and I believe there were
some fundamental problems with the design that we never addressed.
As I recall, this is because we adopted the policy that other projects
should never use config opts they didn't define, so there's no real
need for this feature now.

I'm proposing that we deprecate this for only one cycle and remove it
in U. While this doesn't fit our usual deprecation policy, the fact
that the module has no known users makes me think it's not worth
waiting the full two cycles we normally would.

Change-Id: I04baae695a7ba0bb93c91ea642c4e33c04bf87bd
0: http://codesearch.openstack.org/?q=ConfigFilter&i=nope&files=&repos=