Cap flake8 to be < 3.8.0
The 3.8.0 version introduces some checks that cause failures with the
current code. To avoid that, this caps flake8 to a version that had been
tested withouth errors so hacking does not pull in a newer version.

We will likely want to raise hacking to 3.x and drop this cap, but to
make it easy to backport to stable branches, taking this approach first.

Change-Id: I2a539263e8cfc7d15243d78e20b70b8ed763b8fd
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>