Disable a config test if ran under the root user
The `test_conf_file_permission_denied` is designed to test
if a config file can't be read due to missing file permissions.

However if we run unit tests with the root user we will beneficiat
from the root privileges and so this test will fails. This can
be reproduced by running `tox` with `sudo`.

These changes propose to skip this test if root privileges are
detected.

Closes-bug: #1983358

Change-Id: Idd124082265032c30f83242b612d002d7093aa61